Assignment2/SaliencyPrediction/requirements.txt
timm==0.6.13
transformers==4.27.4
fairscale==0.4.13
huggingface_hub==0.13.4
clip @ git+https://github.com/openai/CLIP.git

pandas~=2.0.3
scipy~=1.10.1
numpy~=1.24.4
matplotlib~=3.7.3
torch~=2.1.0
tensorboardX~=2.6.2.2
PyYAML~=6.0.1
Pillow~=10.1.0
torchvision~=0.16.0
clip~=1.0
accelerate~=0.16.0
diffusers~=0.16.0
datasets~=2.11.0
packaging~=23.2
tqdm~=4.66.1
requests~=2.31.0
scikit-learn~=1.3.2